izvorniKod/backend/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
psycopg2==2.8.6
coreapi==2.3.3
pyyaml==5.3.1
